descry/requirements.txt
Flask==1.0.2
python-sane==2.8.2
Pillow==7.2.0
PyYAML==3.13
pytest==5.3.1
python-dotenv==0.13.0


